Refactoring and fixing issues with NBT reflection
Also added a utility class to help with converting Bukkit EntityTypes to NMS Entity Types. This is for the White Monster Egg Version Material. I may add support for different monster eggs but for now this one is based on the ghast egg. The mechanism of cancelling the spawn is not implemented and is intended to be implemented by plugins that use this API.